ECO-3120: Moved persisnetce mapper to Propel folder, added specification for module constants.
